% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_temp-airr.R
\name{airr_subgroup_abundance}
\alias{airr_subgroup_abundance}
\title{Compare abundance of immune receptor subgroups, such as samples, cell types and disease-specific receptors}
\usage{
airr_subgroup_abundance(idata, method = c("sum", "mean", "median"), by)
}
\description{
Abundance comparing - compare over-representation of immune repertoire subgroups, such as cell clusters or disease-specific receptors
}
