% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-engine-registry-class.R
\name{BackendRegistry}
\alias{BackendRegistry}
\title{Backend Engine Registry}
\description{
Backend Engine Registry

Backend Engine Registry
}
\details{
A registry for backend backends used in data processing. This class allows for
the registration and retrieval of backend engine instances, ensuring that only
valid backends are used within the application.

Initializes the BackendRegistry object.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BackendRegistry-new}{\code{BackendRegistry$new()}}
\item \href{#method-BackendRegistry-names}{\code{BackendRegistry$names()}}
\item \href{#method-BackendRegistry-register}{\code{BackendRegistry$register()}}
\item \href{#method-BackendRegistry-get}{\code{BackendRegistry$get()}}
\item \href{#method-BackendRegistry-clone}{\code{BackendRegistry$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BackendRegistry-new"></a>}}
\if{latex}{\out{\hypertarget{method-BackendRegistry-new}{}}}
\subsection{Method \code{new()}}{
Object Constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BackendRegistry$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BackendRegistry-names"></a>}}
\if{latex}{\out{\hypertarget{method-BackendRegistry-names}{}}}
\subsection{Method \code{names()}}{
Get a list of engine names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BackendRegistry$names()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character vector of engine names.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BackendRegistry-register"></a>}}
\if{latex}{\out{\hypertarget{method-BackendRegistry-register}{}}}
\subsection{Method \code{register()}}{
Register a new backend engine.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BackendRegistry$register(.name, .backend)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.name}}{The name of the backend engine to register. Must be a non-empty string.
If provided a vector, use every elements as an alias.}

\item{\code{.backend}}{An instance of the backend engine to register. Must be a child instance of an R6 class "Backend"}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible. The function is called for its side effect of registering an engine.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BackendRegistry-get"></a>}}
\if{latex}{\out{\hypertarget{method-BackendRegistry-get}{}}}
\subsection{Method \code{get()}}{
Retrieve a registered backend engine by name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BackendRegistry$get(.name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.name}}{The name of the backend engine to retrieve. Must be a non-empty string.}

\item{\code{...}}{Other parameters you pass to the constructor of Backend coded by ".name"}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of the backend engine.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BackendRegistry-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BackendRegistry-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BackendRegistry$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
