% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations_compute_collect.R
\name{compute.ImmunData}
\alias{compute.ImmunData}
\title{Compute ImmunData annotations}
\usage{
\method{compute}{ImmunData}(x, ...)
}
\arguments{
\item{x}{ImmunData object.}

\item{...}{Additional arguments passed to \code{\link[dplyr:compute]{dplyr::compute()}} for
\code{x$annotations}.}
}
\value{
A new \code{ImmunData} object with computed annotations.
}
\description{
Materializes the annotation table of an \code{ImmunData} object via
\code{\link[dplyr:compute]{dplyr::compute()}} and returns a new \code{ImmunData}.
}
\concept{operations}
