% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immundata.R
\name{DataTableImmunData}
\alias{DataTableImmunData}
\title{DataTableImmunData: ImmunData Representation Using tidytable}
\value{
An object of class \code{DataTableImmunData}.
}
\description{
DataTableImmunData: ImmunData Representation Using tidytable

DataTableImmunData: ImmunData Representation Using tidytable
}
\details{
This class extends \code{ImmunData} and represents immune receptor repertoire data
using \code{tidytable} as the internal data structure.
}
\section{Methods}{

\describe{
\item{\code{new(.dataset, .metadata, .repertoire_model = NULL, .clonotype_model = NULL)}}{
Creates a new \code{DataTableImmunData} object.
}
}
}

\section{Super class}{
\code{\link[immundata:ImmunData]{immundata::ImmunData}} -> \code{DataTableImmunData}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataTableImmunData-new}{\code{DataTableImmunData$new()}}
\item \href{#method-DataTableImmunData-clone}{\code{DataTableImmunData$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="immundata" data-topic="ImmunData" data-id="[["><a href='../../immundata/html/ImmunData.html#method-ImmunData-[['><code>immundata::ImmunData$[[()</code></a></span></li>
<li><span class="pkg-link" data-pkg="immundata" data-topic="ImmunData" data-id="arrange"><a href='../../immundata/html/ImmunData.html#method-ImmunData-arrange'><code>immundata::ImmunData$arrange()</code></a></span></li>
<li><span class="pkg-link" data-pkg="immundata" data-topic="ImmunData" data-id="collect"><a href='../../immundata/html/ImmunData.html#method-ImmunData-collect'><code>immundata::ImmunData$collect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="immundata" data-topic="ImmunData" data-id="compute"><a href='../../immundata/html/ImmunData.html#method-ImmunData-compute'><code>immundata::ImmunData$compute()</code></a></span></li>
<li><span class="pkg-link" data-pkg="immundata" data-topic="ImmunData" data-id="count"><a href='../../immundata/html/ImmunData.html#method-ImmunData-count'><code>immundata::ImmunData$count()</code></a></span></li>
<li><span class="pkg-link" data-pkg="immundata" data-topic="ImmunData" data-id="data"><a href='../../immundata/html/ImmunData.html#method-ImmunData-data'><code>immundata::ImmunData$data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="immundata" data-topic="ImmunData" data-id="filter"><a href='../../immundata/html/ImmunData.html#method-ImmunData-filter'><code>immundata::ImmunData$filter()</code></a></span></li>
<li><span class="pkg-link" data-pkg="immundata" data-topic="ImmunData" data-id="group_by"><a href='../../immundata/html/ImmunData.html#method-ImmunData-group_by'><code>immundata::ImmunData$group_by()</code></a></span></li>
<li><span class="pkg-link" data-pkg="immundata" data-topic="ImmunData" data-id="mutate"><a href='../../immundata/html/ImmunData.html#method-ImmunData-mutate'><code>immundata::ImmunData$mutate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="immundata" data-topic="ImmunData" data-id="print"><a href='../../immundata/html/ImmunData.html#method-ImmunData-print'><code>immundata::ImmunData$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="immundata" data-topic="ImmunData" data-id="select"><a href='../../immundata/html/ImmunData.html#method-ImmunData-select'><code>immundata::ImmunData$select()</code></a></span></li>
<li><span class="pkg-link" data-pkg="immundata" data-topic="ImmunData" data-id="slice"><a href='../../immundata/html/ImmunData.html#method-ImmunData-slice'><code>immundata::ImmunData$slice()</code></a></span></li>
<li><span class="pkg-link" data-pkg="immundata" data-topic="ImmunData" data-id="slice_head"><a href='../../immundata/html/ImmunData.html#method-ImmunData-slice_head'><code>immundata::ImmunData$slice_head()</code></a></span></li>
<li><span class="pkg-link" data-pkg="immundata" data-topic="ImmunData" data-id="slice_tail"><a href='../../immundata/html/ImmunData.html#method-ImmunData-slice_tail'><code>immundata::ImmunData$slice_tail()</code></a></span></li>
<li><span class="pkg-link" data-pkg="immundata" data-topic="ImmunData" data-id="summarise"><a href='../../immundata/html/ImmunData.html#method-ImmunData-summarise'><code>immundata::ImmunData$summarise()</code></a></span></li>
<li><span class="pkg-link" data-pkg="immundata" data-topic="ImmunData" data-id="to_list"><a href='../../immundata/html/ImmunData.html#method-ImmunData-to_list'><code>immundata::ImmunData$to_list()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTableImmunData-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataTableImmunData-new}{}}}
\subsection{Method \code{new()}}{
Initializes a \code{DataTableImmunData} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTableImmunData$new(
  .dataset,
  .metadata,
  .repertoire_model = NULL,
  .clonotype_model = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.dataset}}{A data frame containing immune receptor data.}

\item{\code{.metadata}}{A data frame containing metadata information.}

\item{\code{.repertoire_model}}{Optional. A model for repertoire-level analysis.}

\item{\code{.clonotype_model}}{Optional. A model for clonotype-level analysis.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataTableImmunData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataTableImmunData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTableImmunData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
