% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations_dimnames.R
\name{dimnames<-.ImmunData}
\alias{dimnames<-.ImmunData}
\title{Prevent Renaming ImmunData via dimnames}
\usage{
\method{dimnames}{ImmunData}(x) <- value
}
\arguments{
\item{x}{ImmunData object.}

\item{value}{Not used.}
}
\description{
Disallows replacing dimension names on \code{ImmunData}.
}
\concept{operations}
