% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations_dimnames.R
\name{dimnames.ImmunData}
\alias{dimnames.ImmunData}
\title{Get Annotation Dimnames from ImmunData}
\usage{
\method{dimnames}{ImmunData}(x)
}
\arguments{
\item{x}{ImmunData object.}
}
\value{
A list with \code{NULL} row names and annotation column names.
}
\description{
Returns dimension names for an \code{ImmunData} object so that
\code{colnames(idata)} maps to annotation column names.
}
\concept{operations}
