% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations_filter.R
\name{filter_cells}
\alias{filter_cells}
\alias{filter_receptors}
\title{Filter ImmunData by Passed Cell IDs or Barcodes}
\usage{
filter_cells(idata, cells, keep_repertoires = TRUE)

filter_receptors(idata, receptors, keep_repertoires = TRUE)
}
\arguments{
\item{idata}{\link{ImmunData} object.}

\item{cells}{Vector of cell identifiers or barcodes to filter by.}
}
\description{
A short description...
}
\concept{Filtering}
