% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immundata_operations_filter.R
\name{filter_annotations}
\alias{filter_annotations}
\title{Filter ImmunData by Annotations}
\usage{
filter_annotations(idata, ...)
}
\arguments{
\item{idata}{\link{ImmunData} object.}

\item{...}{Parameters passed to \link[dplyr:filter]{dplyr::filter}.}
}
\description{
A short description...
}
