% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immundata_operations_filter.R
\name{filter_barcodes}
\alias{filter_barcodes}
\title{Filter ImmunData by Passed Barcodes}
\usage{
filter_barcodes(idata, barcodes)
}
\arguments{
\item{idata}{\link{ImmunData} object.}

\item{barcodes}{Vector of barcodes to filter by.}
}
\description{
A short description...
}
