% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_available_backends}
\alias{get_available_backends}
\title{Get Available Backends for ImmunData}
\usage{
get_available_backends()
}
\value{
A character vector of registered backend names.
If no backends are registered, returns an empty character vector.
}
\description{
Returns a list of currently registered backends for \code{ImmunData}.
These backends determine the internal data representation,
such as \code{data.frame}, \code{data.table}, or other formats.
}
