% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{imd_drop_cols}
\alias{imd_drop_cols}
\title{Return a list of columns to drop from the dataset}
\usage{
imd_drop_cols(format = "airr")
}
\arguments{
\item{format}{File format}
}
\description{
Return a list of columns to drop from the dataset
}
