% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{imd_files}
\alias{imd_files}
\title{Get Immundata Default File Names}
\usage{
imd_files()
}
\value{
A named list of file names (e.g., \code{receptors.parquet}, \code{annotations.parquet}).
}
\description{
Returns the standardized default filenames for storing receptor-level and annotation-level
data as used in \code{load_repertoires()} and related Immundata I/O functions.
}
