% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{imd_rename_cols}
\alias{imd_rename_cols}
\title{Return a list of columns to rename}
\usage{
imd_rename_cols(format = "default")
}
\arguments{
\item{format}{File format}
}
\description{
Return a list of columns to rename
}
