% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{imd_repertoire_schema}
\alias{imd_repertoire_schema}
\title{Return a typical repertoire columns, which define a repertoire for a specific format}
\usage{
imd_repertoire_schema(format = "airr")
}
\arguments{
\item{format}{File format}
}
\description{
Return a typical repertoire columns, which define a repertoire for a specific format
}
