% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immundata-package.R
\docType{package}
\name{immundata-package}
\alias{immundata}
\alias{immundata-package}
\title{immundata: A Unified Data Layer for Large-Scale Single-Cell, Spatial and Bulk Immunomics}
\description{
Provides a unified data layer for single-cell, spatial and bulk T-cell and B-cell immune receptor repertoire data. Think AnnData or SeuratObject, but for AIRR data, a.k.a. Adaptive Immune Receptor Repertoire, VDJ-seq, RepSeq, or VDJ sequencing data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://immunomind.github.io/docs/}
  \item \url{https://github.com/immunomind/immundata}
  \item Report bugs at \url{https://github.com/immunomind/immundata/issues}
}

}
\author{
\strong{Maintainer}: Vadim I. Nazarov \email{support@immunomind.com} (\href{https://orcid.org/0000-0003-3659-2709}{ORCID})

}
\keyword{internal}
