% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_repertoires.R
\name{ImmunDataLoader}
\alias{ImmunDataLoader}
\title{ImmunDataLoader}
\description{
ImmunDataLoader

ImmunDataLoader
}
\details{
A class for loading immunological data. This loader supports different backend
backends for data processing and can handle multiple data formats.

Initializes a new ImmunDataLoader object with a specified source path and backend backend.

Lists all data files available in the source directory, optionally filtered by format.

Loads a specific data file by name using the configured backend.

Loads all available data files using the configured backend backend.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ImmunDataLoader-new}{\code{ImmunDataLoader$new()}}
\item \href{#method-ImmunDataLoader-list}{\code{ImmunDataLoader$list()}}
\item \href{#method-ImmunDataLoader-load}{\code{ImmunDataLoader$load()}}
\item \href{#method-ImmunDataLoader-load_all}{\code{ImmunDataLoader$load_all()}}
\item \href{#method-ImmunDataLoader-clone}{\code{ImmunDataLoader$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImmunDataLoader-new"></a>}}
\if{latex}{\out{\hypertarget{method-ImmunDataLoader-new}{}}}
\subsection{Method \code{new()}}{
Constructor for ImmunDataLoader
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImmunDataLoader$new(
  .source,
  .backend = "df",
  .recursive = TRUE,
  .cache = FALSE,
  .format = "auto"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.source}}{Character or object. If character, then either a path to a file(s) or a folder(s) with data files.
If object, then either a data frame or a list of data frames.}

\item{\code{.backend}}{Character or Backend; the backend to use for data processing. Default is "df".}

\item{\code{.recursive}}{...}

\item{\code{.cache}}{...}

\item{\code{.format}}{Character; the format of the data files to load. This parameter is optional and defaults to auto" -
automatic detection of formats.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImmunDataLoader-list"></a>}}
\if{latex}{\out{\hypertarget{method-ImmunDataLoader-list}{}}}
\subsection{Method \code{list()}}{
List Data Files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImmunDataLoader$list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character vector of the file names available for loading.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImmunDataLoader-load"></a>}}
\if{latex}{\out{\hypertarget{method-ImmunDataLoader-load}{}}}
\subsection{Method \code{load()}}{
Load a Specific Data File
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImmunDataLoader$load(.filename)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.filename}}{Character; the name of the file to load. Could be a vector if you want to load several files.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data frame or other structure containing the loaded data, depending on thebackend backend.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImmunDataLoader-load_all"></a>}}
\if{latex}{\out{\hypertarget{method-ImmunDataLoader-load_all}{}}}
\subsection{Method \code{load_all()}}{
Load All Data Files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImmunDataLoader$load_all()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of data frames or other structures containing the loaded data, depending on the backend backend.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImmunDataLoader-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ImmunDataLoader-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImmunDataLoader$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
