% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{immuntest}
\alias{immuntest}
\title{Run a Basic Test for ImmunDataLoader}
\usage{
immuntest(.backend = "df")
}
\arguments{
\item{.backend}{Character string specifying the backend to use.
Supported backends include \code{"df"} (data frame) and others defined in \code{ImmunDataLoader}.}
}
\value{
An \code{ImmunData} object containing the loaded dataset.
}
\description{
This function initializes an \code{ImmunDataLoader} and loads an example dataset.
}
