% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations_dimnames.R
\name{names<-.ImmunData}
\alias{names<-.ImmunData}
\title{Prevent Renaming ImmunData via names}
\usage{
\method{names}{ImmunData}(x) <- value
}
\arguments{
\item{x}{ImmunData object.}

\item{value}{Not used.}
}
\description{
Disallows replacing names on \code{ImmunData}.
}
\concept{operations}
