% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_temp-receptor.R
\name{receptor_public_discovery}
\alias{receptor_public_discovery}
\title{Discovery of public immune receptors}
\usage{
receptor_public_discovery(
  idata,
  method = c("incidence", "fischer", "wilcox"),
  by = NULL
)
}
\value{
Annotated \link{ImmunData} with either a number of repertoires this receptor found, or a corrected p-value
for Fischer or Wilcox-Mann tests.
}
\description{
A short description...
}
