% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_temp-receptor.R
\name{receptor_query}
\alias{receptor_query}
\title{Database query for annotating disease-specific immune receptors}
\usage{
receptor_query(idata, db, idata_col, db_col, method, max_dist)
}
\description{
A short description...
}
