% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_temp-receptor.R
\name{receptor_track}
\alias{receptor_track}
\title{Tracking a list of target immune receptors across repertoires}
\usage{
receptor_track(
  idata,
  targets,
  idata_col,
  targets_col,
  method = c("exact", "hamm", "lev"),
  max_dist = 2
)
}
\description{
Tracking a list of target immune receptors across repertoires
}
\section{-- TRACK --}{
If no seq_options passed than seq_options is created automatically and takes the first column from the targets
\enumerate{
\item \link{receptor_track} Find receptors of interest
1a) List of CDR3 - vector + col to match (screw that, just pass a data frame)
1b) List of CDR3 + other columns
1c) List of CDR3 alpha + CDR3 beta (+ other columns)
2a) Find abundance of found receptors in an ordered repertoires
2b) Find abundance of receptors close by distance in an ordered repertoires
3a) Visualise via alluvial
3b) Compute PCA on abundances and visualise
}

Example: visualise the size of cell clusters those receptors belong to across different therapy time points
}

