% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_annotations_write.R
\name{write_annotations}
\alias{write_annotations}
\title{Write immune receptor annotations to single-cell object metadata - Seurat, AnnData, etc.}
\usage{
write_annotations(idata, target, cells = NA)
}
\description{
Write immune receptor annotations to single-cell object metadata - Seurat, AnnData, etc.
}
