% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immundata_operations_agg.R
\name{agg_repertoires}
\alias{agg_repertoires}
\title{Define repertoires by aggregating annotation fields for ImmunData}
\usage{
agg_repertoires(idata, schema = "repertoire_id", sep = "-")
}
\arguments{
\item{idata}{\link{ImmunData} object.}

\item{schema}{A character vector of column names in \code{.annotations} used to define repertoire grouping.}

\item{sep}{Currently unused. A string separator used to concatenate multiple columns into a single repertoire ID.}
}
\description{
Aggregates receptors to repertoires by selected annotating columns.
}
