% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations_compute_collect.R
\name{collect.ImmunData}
\alias{collect.ImmunData}
\title{Collect ImmunData annotations}
\usage{
\method{collect}{ImmunData}(x, ...)
}
\arguments{
\item{x}{ImmunData object.}

\item{...}{Additional arguments passed to \code{\link[dplyr:collect]{dplyr::collect()}} for
\code{x$annotations}.}
}
\value{
A tibble with collected annotations.
}
\description{
Collects annotations from an \code{ImmunData} object and returns them as a tibble.
Factor columns are converted to character.
}
\concept{operations}
