% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immundata.R
\name{collect.ImmunData}
\alias{collect.ImmunData}
\title{Collects all lazy tables in the \code{ImmunData} object into memory using \code{dplyr::collect()}.}
\usage{
\method{collect}{ImmunData}(immdata, ...)
}
\arguments{
\item{immdata}{An \code{ImmunData} object.}

\item{...}{Other arguments passed to \link[dplyr:compute]{dplyr::collect}}
}
\value{
A new \code{ImmunData} object with in-memory receptor and annotation tables.
}
\description{
Collects all lazy tables in the \code{ImmunData} object into memory using \code{dplyr::collect()}.
}
