% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immundata.R
\name{compute.ImmunData}
\alias{compute.ImmunData}
\title{Materializes all lazy tables in the \code{ImmunData} object using \code{dplyr::compute()}.}
\usage{
\method{compute}{ImmunData}(immdata, ...)
}
\arguments{
\item{immdata}{An \code{ImmunData} object.}

\item{...}{Other arguments passed to \link[dplyr:compute]{dplyr::compute}}
}
\value{
A new \code{ImmunData} object with computed receptor and annotation tables.
}
\description{
Materializes all lazy tables in the \code{ImmunData} object using \code{dplyr::compute()}.
}
