% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations_count.R
\name{count.ImmunData}
\alias{count.ImmunData}
\title{Count the number of chains in ImmunData}
\usage{
\method{count}{ImmunData}(idata)
}
\arguments{
\item{idata}{ImmunData object.}
}
\description{
Count the number of chains in ImmunData
}
