% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations_count.R
\name{count.ImmunData}
\alias{count.ImmunData}
\title{Count the number of chains in ImmunData}
\usage{
\method{count}{ImmunData}(x, ..., wt = NULL, sort = FALSE, name = NULL)
}
\arguments{
\item{x}{ImmunData object.}

\item{...}{Not used.}

\item{wt}{Not used.}

\item{sort}{Not used.}

\item{name}{Not used.}
}
\description{
Count the number of chains in ImmunData
}
\concept{operations}
