% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immundata_operations_filter.R
\name{filter_receptors}
\alias{filter_receptors}
\title{Filter ImmunData by Receptor Features}
\usage{
filter_receptors(idata, ...)
}
\arguments{
\item{idata}{\link{ImmunData} object.}

\item{...}{Parameters passed to \link[dplyr:filter]{dplyr::filter}.}
}
\description{
A short description...
}
