% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_test.R
\name{get_test_immundata}
\alias{get_test_immundata}
\title{Get test datasets from \code{immundata}}
\usage{
get_test_immundata(repertoire_schema = "<auto>")
}
\description{
Get test datasets from \code{immundata}
}
\keyword{internal}
