% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{imd_schema}
\alias{imd_schema}
\alias{imd_schema_sym}
\alias{imd_meta_schema}
\alias{imd_files}
\alias{imd_rename_cols}
\alias{imd_drop_cols}
\alias{imd_repertoire_schema}
\title{Get Immundata Internal Schema Field Names}
\usage{
imd_schema(key = NULL)

imd_schema_sym(key = NULL)

imd_meta_schema()

imd_files()

imd_rename_cols(format = "default")

imd_drop_cols(format = "airr")

imd_repertoire_schema(format = "airr")
}
\value{
A named list of schema field names.
}
\description{
Returns the standardized field names used across Immundata objects and processing functions,
as defined in \code{IMD_GLOBALS$schema}. These include column names for cell ids or barcodes, receptors,
repertoires, and related metadata.
}
