% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{imd_schema}
\alias{imd_schema}
\title{Get Immundata Internal Schema Field Names}
\usage{
imd_schema(val = NULL)
}
\value{
A named list of schema field names.
}
\description{
Returns the standardized field names used across Immundata objects and processing functions,
as defined in \code{IMD_GLOBALS$schema}. These include column names for cell ids or barcodes, receptors,
repertoires, and related metadata.
}
