% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immundata-package.R
\docType{package}
\name{immundata-package}
\alias{immundata}
\alias{immundata-package}
\title{immundata: A Unified Data Layer for Single-Cell, Spatial and Bulk Immunomics}
\description{
Provides a unified data layer for single-cell, spatial and bulk T-cell and B-cell immune receptor repertoire data, integrating diverse data formats such as AIRR and raw sequencing files. Includes utilities for receptor aggregation, metadata normalization, and clonotype filtering.
}
\author{
\strong{Maintainer}: Vadim I. Nazarov \email{support@immunomind.com} (\href{https://orcid.org/0000-0003-3659-2709}{ORCID})

}
\keyword{internal}
