% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immundata.R
\name{ImmunData}
\alias{ImmunData}
\title{ImmunData: A Unified receptors Structure for Immune Receptor receptors}
\description{
ImmunData: A Unified receptors Structure for Immune Receptor receptors

ImmunData: A Unified receptors Structure for Immune Receptor receptors
}
\details{
\code{ImmunData} is an abstract class for managing and processing
immune receptor repertoire receptors, supporting flexible backends and
efficient receptors transformations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.receptors}}{description}

\item{\code{.annotations}}{description}

\item{\code{schema_receptor}}{description}

\item{\code{schema_repertoire}}{description}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ImmunData-new}{\code{ImmunData$new()}}
\item \href{#method-ImmunData-build_repertoires}{\code{ImmunData$build_repertoires()}}
\item \href{#method-ImmunData-print}{\code{ImmunData$print()}}
\item \href{#method-ImmunData-filter_data}{\code{ImmunData$filter_data()}}
\item \href{#method-ImmunData-filter_annot}{\code{ImmunData$filter_annot()}}
\item \href{#method-ImmunData-filter_barcodes}{\code{ImmunData$filter_barcodes()}}
\item \href{#method-ImmunData-count}{\code{ImmunData$count()}}
\item \href{#method-ImmunData-[[}{\code{ImmunData$[[()}}
\item \href{#method-ImmunData-clone}{\code{ImmunData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImmunData-new"></a>}}
\if{latex}{\out{\hypertarget{method-ImmunData-new}{}}}
\subsection{Method \code{new()}}{
Initializes an \code{ImmunData} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImmunData$new(receptors, annotations, schema)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{receptors}}{description}

\item{\code{annotations}}{description}

\item{\code{schema}}{description}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImmunData-build_repertoires"></a>}}
\if{latex}{\out{\hypertarget{method-ImmunData-build_repertoires}{}}}
\subsection{Method \code{build_repertoires()}}{
Define how this dataset groups receptors to repertoires.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImmunData$build_repertoires(schema = "repertoire_id", sep = "-")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sep}}{description}

\item{\code{columns}}{description}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImmunData-print"></a>}}
\if{latex}{\out{\hypertarget{method-ImmunData-print}{}}}
\subsection{Method \code{print()}}{
Prints the class information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImmunData$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImmunData-filter_data"></a>}}
\if{latex}{\out{\hypertarget{method-ImmunData-filter_data}{}}}
\subsection{Method \code{filter_data()}}{
Filters rows based on conditions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImmunData$filter_data(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Filter conditions.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{ImmunData} object with filtered rows.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImmunData-filter_annot"></a>}}
\if{latex}{\out{\hypertarget{method-ImmunData-filter_annot}{}}}
\subsection{Method \code{filter_annot()}}{
Filters rows based on conditions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImmunData$filter_annot(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImmunData-filter_barcodes"></a>}}
\if{latex}{\out{\hypertarget{method-ImmunData-filter_barcodes}{}}}
\subsection{Method \code{filter_barcodes()}}{
Filters rows based on conditions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImmunData$filter_barcodes(barcodes = c())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{barcodes}}{description}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImmunData-count"></a>}}
\if{latex}{\out{\hypertarget{method-ImmunData-count}{}}}
\subsection{Method \code{count()}}{
Counts occurrences of groupings.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImmunData$count(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Variables to count.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{ImmunData} object with count summary.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImmunData-[["></a>}}
\if{latex}{\out{\hypertarget{method-ImmunData-[[}{}}}
\subsection{Method \code{[[()}}{
Extracts sample-specific receptors.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImmunData$[[(.sample)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.sample}}{The sample identifier.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A filtered dataset containing only the selected sample.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImmunData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ImmunData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImmunData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
