% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_repertoires_processing.R
\name{make_default_preprocessing}
\alias{make_default_preprocessing}
\alias{make_default_postprocessing}
\alias{make_exclude_columns}
\alias{make_productive_filter}
\alias{make_barcode_prefix}
\title{Preprocessing and postprocessing of input immune repertoire files}
\usage{
make_default_preprocessing(format = c("airr", "10x"))

make_default_postprocessing()

make_exclude_columns(cols = imd_drop_cols("airr"))

make_productive_filter(col_name = c("productive"), truthy = TRUE)

make_barcode_prefix(prefix_col = "Prefix")
}
\description{
Preprocessing and postprocessing of input immune repertoire files
}
\concept{Preprocessing and postprocessing}
